+++ bench/Main.hs
@@ -10,10 +10,7 @@ import           Data.Monoid             ((<>))
 import           Data.String.Interpolate (i)
 import           Data.Text               (Text)
 import qualified Data.Text               as T
-import qualified Data.Text.Format        as TF
-import           Data.Text.Format.Params as TF
-import qualified Data.Text.Lazy          as LT
-import           Fmt                     ((+|), (|+))
+import           Fmt                     ((+|), (|+), format)
 import           Formatting              (Format, formatToString, int, sformat, stext,
                                           string, (%))
 import           Text.Printf             (printf)
@@ -24,12 +21,6 @@ import           Criterion.Main          (defaultMain)
 ----------------------------------------------------------------------------
 -- Format utility functions
 ----------------------------------------------------------------------------
-
-format' :: TF.Params ps => TF.Format -> ps -> Text
-format' f = LT.toStrict . TF.format f
-
-formatS :: TF.Params ps => TF.Format -> ps -> String
-formatS f = LT.unpack . TF.format f
 
 -- Shorter alias for @formatToString@.
 fs :: Format String a -> a
@@ -69,10 +60,10 @@ main = defaultMain
     [ bTextGroup (1 :: Int, 2 :: Int)
       [ taggedB "fmt" $
           \(a,b) -> "hello "+|a|+" world "+|b|+""
+      , taggedB "fmt old-style" $
+          \(a,b) -> format "hello {} world {}" a b :: Text
       , taggedB "formatting" $
           \(a,b) -> sformat ("hello "%int%" world "%int) a b
-      , taggedB "text-format" $
-          format' "hello {} world {}"
       , taggedB "interpolate" $
           \(a,b) -> T.pack [i|hello #{a} world #{b}|]
       , taggedB "show" $
@@ -83,10 +74,10 @@ main = defaultMain
     , bStringGroup (1 :: Int, 2 :: Int)
       [ taggedB "fmt" $
           \(a,b) -> "hello "+|a|+" world "+|b|+""
+      , taggedB "fmt old-style" $
+          \(a,b) -> format "hello {} world {}" a b :: String
       , taggedB "formatting" $
           \(a,b) -> fs ("hello "%int%" world "%int) a b
-      , taggedB "text-format" $
-          formatS "hello {} world {}"
       , taggedB "interpolate" $
           \(a,b) -> [i|hello #{a} world #{b}|]
       , taggedB "show" $
@@ -100,10 +91,10 @@ main = defaultMain
     [ bTextGroup (9 :: Int, "Beijing" :: Text)
       [ taggedB "fmt" $
           \(n,city) -> "There are "+|n|+" million bicycles in "+|city|+"."
+      , taggedB "fmt old-style" $
+          \(n,city) -> format "There are {} million bicycles in {}." n city :: Text
       , taggedB "formatting" $
           \(n,city) -> sformat ("There are "%int%" million bicycles in "%stext%".") n city
-      , taggedB "text-format" $
-          format' "There are {} million bicycles in {}."
       , taggedB "interpolate" $
           \(n,city) -> T.pack [i|There are #{n} million bicycles in #{city}.|]
       , taggedB "show" $
@@ -114,10 +105,10 @@ main = defaultMain
     , bStringGroup (9 :: Int, "Beijing" :: String)
       [ taggedB "fmt" $
           \(n,city) -> "There are "+|n|+" million bicycles in "+|city|+"."
+      , taggedB "fmt old-style" $
+          \(n,city) -> format "There are {} million bicycles in {}." n city :: String
       , taggedB "formatting" $
           \(n,city) -> fs ("There are "%int%" million bicycles in "%string%".") n city
-      , taggedB "text-format" $
-          formatS "There are {} million bicycles in {}."
       , taggedB "interpolate" $
           \(n,city) -> [i|There are #{n} million bicycles in #{city}.|]
       , taggedB "show" $