+++ tests/Main.hs
@@ -22,6 +22,8 @@ import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy as BSL
 -- Generics
 import GHC.Generics
+-- Call stacks
+import Data.CallStack
 -- Tests
 import Test.Hspec
 
@@ -149,8 +151,6 @@ test_floatingPoint = describe "floating-point" $ do
     exptF 2 f1_3 ==#> "1.30e0"
   it "fixedF" $ do
     fixedF 2 f1_3 ==#> "1.30"
-  it "precF" $ do
-    precF 2 f1_3 ==#> "1.3"
 
 test_hex :: Spec
 test_hex = describe "'hexF'" $ do
@@ -508,14 +508,14 @@ test_jsonMapF = describe "'jsonMapF'" $ do
 test_tuples :: Spec
 test_tuples = describe "tuples" $ do
   test_tupleSimple
-  describe "tupleLikeF" $ do
+  describe "tupleF on lists" $ do
     test_tupleOneLine
     test_tupleMultiline
 
 test_tupleSimple :: Spec
 test_tupleSimple = it "tupleF" $ do
   -- we don't need complicated tests here, they're all tested in
-  -- 'tupleLikeF' tests
+  -- "tupleF on lists" tests
   tupleF (n, s) ==#> "(25, !)"
   tupleF (n, s, n, s) ==#> "(25, !, 25, !)"
   tupleF (n, s, n, s, 'a', 'b', 'c', 'd') ==#>
@@ -524,21 +524,21 @@ test_tupleSimple = it "tupleF" $ do
 test_tupleOneLine :: Spec
 test_tupleOneLine = describe "one-line" $ do
   it "()" $ do
-    tupleLikeF [] ==#> "()"
+    tupleF ([] :: [Builder]) ==#> "()"
   it "('')" $ do
-    tupleLikeF [""] ==#> "()"
+    tupleF ([""] :: [Builder]) ==#> "()"
   it "(a)" $ do
-    tupleLikeF ["a"] ==#> "(a)"
+    tupleF (["a"] :: [Builder]) ==#> "(a)"
   it "(a,b)" $ do
-    tupleLikeF ["a", "b"] ==#> "(a, b)"
+    tupleF (["a", "b"] :: [Builder]) ==#> "(a, b)"
   it "(a,'')" $ do
-    tupleLikeF ["a", ""] ==#> "(a, )"
+    tupleF (["a", ""] :: [Builder]) ==#> "(a, )"
   it "('',b)" $ do
-    tupleLikeF ["", "b"] ==#> "(, b)"
+    tupleF (["", "b"] :: [Builder]) ==#> "(, b)"
   it "('','')" $ do
-    tupleLikeF ["", ""] ==#> "(, )"
+    tupleF (["", ""] :: [Builder]) ==#> "(, )"
   it "(a,b,c)" $ do
-    tupleLikeF ["a", "ba", "caba"] ==#> "(a, ba, caba)"
+    tupleF (["a", "ba", "caba"] :: [Builder]) ==#> "(a, ba, caba)"
 
 test_tupleMultiline :: Spec
 test_tupleMultiline = describe "multiline" $ do
@@ -546,45 +546,45 @@ test_tupleMultiline = describe "multiline" $ do
   someEmpty
   it "weird case" $ do
     -- not sure whether I should fix it or not
-    tupleLikeF ["a\n"] ==#> "(a\n)"
+    tupleF ["a\n" :: Builder] ==#> "(a\n)"
 
   where
     allNonEmpty = describe "all non-empty" $ do
       it "1 element (2 lines)" $ do
-          tupleLikeF ["a\nx"] ==#> [text|
+          tupleF ["a\nx" :: Builder] ==#> [text|
             ( a
               x )
             |]
-          tupleLikeF ["a\n x"] ==#> [text|
+          tupleF ["a\n x" :: Builder] ==#> [text|
             ( a
                x )
             |]
-          tupleLikeF [" a\nx\n"] ==#> [text|
+          tupleF [" a\nx\n" :: Builder] ==#> [text|
             (  a
               x )
             |]
       it "1 element (3 lines)" $ do
-          tupleLikeF ["a\nb\nc"] ==#> [text|
+          tupleF ["a\nb\nc" :: Builder] ==#> [text|
             ( a
               b
               c )
             |]
       it "2 elements (1 line + 2 lines)" $ do
-          tupleLikeF ["a", "b\nc"] ==#> [text|
+          tupleF ["a", "b\nc" :: Builder] ==#> [text|
             ( a
             ,
               b
               c )
             |]
       it "2 elements (2 lines + 1 line)" $ do
-          tupleLikeF ["a\nb", "c"] ==#> [text|
+          tupleF ["a\nb", "c" :: Builder] ==#> [text|
             ( a
               b
             ,
               c )
             |]
       it "3 elements (each has 2 lines)" $ do
-          tupleLikeF ["a\nb", "c\nd", "e\nf"] ==#> [text|
+          tupleF ["a\nb", "c\nd", "e\nf" :: Builder] ==#> [text|
             ( a
               b
             ,
@@ -597,21 +597,21 @@ test_tupleMultiline = describe "multiline" $ do
 
     someEmpty = describe "some empty" $ do
       it "2 elements (0 + 2)" $ do
-          tupleLikeF ["", "a\nb"] ==#> [text|
+          tupleF ["", "a\nb" :: Builder] ==#> [text|
             (
             ,
               a
               b )
             |]
       it "2 elements (2 + 0)" $ do
-          tupleLikeF ["a\nb", ""] ==#> [text|
+          tupleF ["a\nb", "" :: Builder] ==#> [text|
             ( a
               b
             ,
               )
             |]
       it "3 elements (0 + 2 + 0)" $ do
-          tupleLikeF ["", "a\nb", ""] ==#> [text|
+          tupleF ["", "a\nb", "" :: Builder] ==#> [text|
             (
             ,
               a
@@ -620,7 +620,7 @@ test_tupleMultiline = describe "multiline" $ do
               )
             |]
       it "3 elements (2 + 0 + 2)" $ do
-          tupleLikeF ["a\nb", "", "c\nd"] ==#> [text|
+          tupleF ["a\nb", "", "c\nd" :: Builder] ==#> [text|
             ( a
               b
             ,
@@ -629,7 +629,7 @@ test_tupleMultiline = describe "multiline" $ do
               d )
             |]
       it "4 elements (2 + 0 + 0 + 2)" $ do
-          tupleLikeF ["a\nb", "", "", "c\nd"] ==#> [text|
+          tupleF ["a\nb", "", "", "c\nd" :: Builder] ==#> [text|
             ( a
               b
             ,
@@ -706,8 +706,8 @@ test_generic = describe "'genericF'" $ do
 -- Utilities
 ----------------------------------------------------------------------------
 
-(==%>) :: Text -> Text -> Expectation
+(==%>) :: HasCallStack => Text -> Text -> Expectation
 (==%>) = shouldBe
 
-(==#>) :: Builder -> Text -> Expectation
+(==#>) :: HasCallStack => Builder -> Text -> Expectation
 (==#>) a b = a `shouldBe` build (T.replace "_" " " b)