+++ tests/doctests.hs
@@ -1,0 +1,1 @@
+{-# OPTIONS_GHC -F -pgmF doctest-discover -optF tests/doctest-config.json #-}